Streamline user bundle build process
The annotation processor has been split into a separate (sub)project
since it is a build-time only dependency and does not need to be with
the main SDK code, which contains mostly runtime code.

A new "Airflow SDK plugin" is also added to wrap logic in the user's
build file so the user can simply add the plugin, fill out required
fields in the 'airflowBundle' block, and that's it. This should make
building a Java task bundle a lot easier.
