+++ java-sdk/gradle.properties
@@ -19,4 +19,4 @@ org.gradle.configuration-cache=true
 
 airflowSupervisorSchemaVersion=2026-06-16
 
-sdkVersion=1.0.0-SNAPSHOT
+projectVersion=1.0.0-SNAPSHOT
