+++ java-sdk/plugin/build.gradle.kts
@@ -1,0 +1,56 @@
+/*
+ * Licensed to the Apache Software Foundation (ASF) under one
+ * or more contributor license agreements.  See the NOTICE file
+ * distributed with this work for additional information
+ * regarding copyright ownership.  The ASF licenses this file
+ * to you under the Apache License, Version 2.0 (the
+ * "License"); you may not use this file except in compliance
+ * with the License.  You may obtain a copy of the License at
+ *
+ *   http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing,
+ * software distributed under the License is distributed on an
+ * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
+ * KIND, either express or implied.  See the License for the
+ * specific language governing permissions and limitations
+ * under the License.
+ */
+
+plugins {
+    `java-gradle-plugin`
+    id("airflow-publish")
+}
+
+repositories {
+    gradlePluginPortal()
+}
+
+dependencies {
+    implementation("com.gradleup.shadow:shadow-gradle-plugin:9.1.0") // Last supporting Java 11.
+}
+
+gradlePlugin {
+    plugins {
+        create("airflowSdk") {
+            id = "org.apache.airflow.sdk"
+            implementationClass = "org.apache.airflow.sdk.plugin.AirflowSdkPlugin"
+            displayName = "Apache Airflow SDK Plugin"
+            description = "Pack a Dag bundle JAR for Airflow to consume."
+        }
+    }
+}
+
+publishing {
+    publications {
+        withType<MavenPublication>().configureEach {
+            if (name == "pluginMaven") {
+                artifactId = "airflow-sdk-gradle-plugin"
+                pom {
+                    name = "Apache Airflow SDK Gradle Plugin"
+                    description = "Gradle plugin for building Apache Airflow Dag bundle JARs."
+                }
+            }
+        }
+    }
+}
