+++ java-sdk/settings.gradle.kts
@@ -1,13 +1,6 @@
-/*
- * This file was generated by the Gradle 'init' task.
- *
- * The settings file is used to specify which projects to include in your build.
- * For more detailed information on multi-project builds, please refer to https://docs.gradle.org/9.2.1/userguide/multi_project_builds.html in the Gradle documentation.
- */
-
 plugins {
     id("org.gradle.toolchains.foojay-resolver-convention").version("0.10.0")
 }
 
 rootProject.name = "airflow-java-sdk"
-include("example", "sdk")
+include("example", "plugin", "processor", "sdk")
